refactor(network): get rid of capsule-builder
