+++ src/api/consumer/lib/capsule-isolate.ts
@@ -1,5 +1,6 @@
 import { loadConsumerIfExist } from '../../../consumer';
-import CapsuleBuilder, { Options } from '../../../extensions/network/capsule-builder';
+import Network, { Options } from '../../../extensions/network/network';
+import capsuleOrchestrator from '../../../extensions/network/orchestrator/orchestrator';
 import { CapsuleOptions } from '../../../extensions/network/orchestrator/types';
 import CapsuleList from '../../../extensions/capsule/capsule-list';
 import { PackageManager } from '../../../extensions/package-manager';
@@ -11,6 +12,9 @@ export default (async function capsuleIsolate(
 ): Promise<CapsuleList> {
   const consumer = await loadConsumerIfExist();
   if (!consumer) throw new Error('no consumer found');
-  const capsuleBuilder = new CapsuleBuilder(consumer.getPath(), new PackageManager('librarian'));
-  return capsuleBuilder.isolateComponents(bitIds, capsuleOptions, options, consumer);
+  await capsuleOrchestrator.buildPools();
+  const network = new Network(capsuleOrchestrator, new PackageManager('librarian'));
+  // const capsuleBuilder = new Network(consumer.getPath(), new PackageManager('librarian'));
+  return network.create(bitIds, capsuleOptions, options, consumer);
+  // return capsuleBuilder.isolateComponents(bitIds, capsuleOptions, options, consumer);
 });
