+++ src/extensions/workspace/workspace.manifest.ts
@@ -1,11 +1,11 @@
 import workspaceProvider from './workspace.provider';
 import { ScopeExt } from '../scope';
 import { ComponentFactoryExt } from '../component';
-import { CapsuleExt } from '../capsule';
+import { NetworkExt } from '../network';
 
 export default {
   name: 'workspace',
-  dependencies: [ScopeExt, ComponentFactoryExt, CapsuleExt],
+  dependencies: [ScopeExt, ComponentFactoryExt, NetworkExt],
   config: {
     /**
      * set the default structure of components in your project.
